scripts/requirements.txt
Pillow==10.2.0
requests==2.31.0
beautifulsoup4==4.12.3 